requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Google Drive API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

